+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+##  [0.2.2] - 2017-01-30
+#### Added
+- Response ```getExtendedError```
+#### Changed
+- YourMembershipClient ```makeCall``` function now throws exception when the Response ```hasError```
+- Response ```getError``` reports the extended error if it available
+
+####
 ##  [0.2.1] - 2017-01-12
 #### Changed
 - Use Container to instantiate GuzzleClient in ServiceProvider